requirements.txt
openai
google-cloud-vision
google-auth
google-generativeai
diffusers
torch